refactor: Update dependencies and prevent runtime error from rusqlite
Sqlite uses i64 for storing integers, so trying to get a u64 out of the
database can fail if the value is outside the valid range for a u64
(e.g., negative). It is safer to store and retrieve u32 if you don't
need all the bits.

+++ src/transaction/core.rs
@@ -218,7 +218,7 @@ pub fn get_transaction(id: TransactionId, connection: &Connection) -> Result<Tra
 ///
 /// # Errors
 /// This function will return a [Error::SqlError] there is some SQL error.
-pub fn count_transactions(connection: &Connection) -> Result<usize, Error> {
+pub fn count_transactions(connection: &Connection) -> Result<u32, Error> {
     connection
         .query_row("SELECT COUNT(id) FROM \"transaction\";", [], |row| {
             row.get(0)
